+++ driver/src/main/com/mongodb/reactivestreams/client/MongoCollection.java
@@ -24,6 +24,7 @@ import com.mongodb.annotations.ThreadSafe;
 import com.mongodb.bulk.BulkWriteResult;
 import com.mongodb.client.model.BulkWriteOptions;
 import com.mongodb.client.model.CountOptions;
+import com.mongodb.client.model.DeleteOptions;
 import com.mongodb.client.model.FindOneAndDeleteOptions;
 import com.mongodb.client.model.FindOneAndReplaceOptions;
 import com.mongodb.client.model.FindOneAndUpdateOptions;
@@ -247,7 +248,6 @@ public interface MongoCollection<TDocument> {
      */
     <TResult> AggregatePublisher<TResult> aggregate(List<? extends Bson> pipeline, Class<TResult> clazz);
 
-
     /**
      * Aggregates documents according to the specified map-reduce function.
      *
@@ -338,12 +338,33 @@ public interface MongoCollection<TDocument> {
     Publisher<DeleteResult> deleteOne(Bson filter);
 
     /**
+     * Removes at most one document from the collection that matches the given filter.  If no documents match, the collection is not
+     * modified.
+     *
+     * @param filter the query filter to apply the the delete operation
+     * @param options the options to apply to the delete operation
+     * @return a publisher with a single element the DeleteResult or with an com.mongodb.MongoException
+     * @since 1.5
+     */
+    Publisher<DeleteResult> deleteOne(Bson filter, DeleteOptions options);
+
+    /**
      * Removes all documents from the collection that match the given query filter.  If no documents match, the collection is not modified.
      *
      * @param filter the query filter to apply the the delete operation
      * @return a publisher with a single element the DeleteResult or with an com.mongodb.MongoException
      */
     Publisher<DeleteResult> deleteMany(Bson filter);
+
+    /**
+     * Removes all documents from the collection that match the given query filter.  If no documents match, the collection is not modified.
+     *
+     * @param filter the query filter to apply the the delete operation
+     * @param options the options to apply to the delete operation
+     * @return a publisher with a single element the DeleteResult or with an com.mongodb.MongoException
+     * @since 1.5
+     */
+    Publisher<DeleteResult> deleteMany(Bson filter, DeleteOptions options);
 
     /**
      * Replace a document in the collection according to the specified arguments.